move io module and fix test case
